Refactor python code structure
